feat(tui): refactor TUIApp and introduce new components for improved state management
- Extracted the StatusBar, ToolCallsPanel, and TransparencyPanel components from TUIApp to enhance modularity and prevent unnecessary re-renders.
- Implemented the useAgentEvents hook to centralize agent event handling and state updates, improving code organization and maintainability.
- Introduced useThrottledState for efficient state updates, reducing re-renders during high-frequency events.
- Updated imports and removed unused code to streamline the application structure.

These changes significantly enhance the performance and maintainability of the TUI application.
