+++ attocode/src/tui/hooks/index.ts
@@ -11,3 +11,13 @@ export {
   type PruneStats,
   type UseMessagePruningResult,
 } from './useMessagePruning.js';
+
+export { useThrottledState, useThrottledCallback } from './use-throttled-state.js';
+
+export {
+  useAgentEvents,
+  type AgentEventSource,
+  type UseAgentEventsRefs,
+  type UseAgentEventsSetters,
+  type UseAgentEventsOptions,
+} from './use-agent-events.js';
